# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Shared sMovie API specification for Admin and Streaming App
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface VersionInfo {
  version: string;
  updatedAt?: number;
}

export interface EpisodeItem {
  id?: string;
  title?: string;
  epNo?: string;
  airDate?: string;
  videoUrl?: string;
  videoFileName?: string;
  subtitles?: string[];
}

export interface SeasonItem {
  seasonNo?: number;
  name?: string;
  episodes?: EpisodeItem[];
}

export interface LibraryItem {
  id: string;
  title: string;
  year?: string;
  contentType?: string;
  media_type?: string;
  language?: string;
  categories?: string[];
  nav_chips?: string[];
  tmdb_id?: string;
  poster_url?: string;
  poster_paths?: string[];
  backdrop_url?: string;
  backdrop_paths?: string[];
  logo_paths?: string[];
  video_url?: string;
  teaser_url?: string;
  quality?: string;
  audio?: string[];
  subtitles?: string[];
  overview?: string;
  rating?: number;
  vote_average?: number;
  seasons?: SeasonItem[];
  owner?: string;
  api_key?: string;
  status?: string;
  addedAt?: number;
  updatedAt?: number;
}

export interface NotificationItem {
  id: string;
  title: string;
  message?: string;
  item_id?: string;
  poster_url?: string;
  timestamp?: number;
}

export type GetLibraryParams = {
category?: string;
contentType?: GetLibraryContentType;
query?: string;
};

export type GetLibraryContentType = typeof GetLibraryContentType[keyof typeof GetLibraryContentType];


export const GetLibraryContentType = {
  all: 'all',
  movie: 'movie',
  series: 'series',
} as const;

export type DeleteLibraryItem200 = {
  success?: boolean;
  id?: string;
};

export type PublishItem200 = {
  success?: boolean;
  item?: LibraryItem;
};