# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Shared sMovie API specification for Admin and Streaming App
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}